# docs/swagger.yaml
basePath: /api/v1
definitions:
  apitypes.MassCalculationJSON:
    properties:
      creator_login:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_form:
        type: string
      id:
        type: integer
      moderator_login:
        type: string
      output_koef:
        type: number
      status:
        type: string
    type: object
...
  apitypes.ReactionJSON:
    properties:
      conversation_factor:
        type: number
      description:
        type: string
      id:
        type: integer
      img_link:
        type: string
      product:
        type: string
      reagent:
        type: string
      title:
        type: string
    type: object
  apitypes.StatusJSON:
    properties:
      status:
        type: string
    type: object
...
  handler.MassCalculationResponse:
    properties:
      calculation:
        $ref: '#/definitions/apitypes.MassCalculationJSON'
      reactions:
        items:
          $ref: '#/definitions/handler.ReactionWithOutput'
        type: array
    type: object
  handler.ReactionWithOutput:
    properties:
      input_mass:
        type: number
      output_mass:
        type: number
      reaction:
        $ref: '#/definitions/apitypes.ReactionJSON'
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для работы с реакциями
  title: Reactions API
  version: "1.0"
paths:
  /mass-calculations:
    get:
      description: Возвращает список расчетов с возможностью фильтрации по диапазону
        дат и статусу.
      parameters:
      - description: Нижняя граница даты (YYYY-MM-DD)
        in: query
        name: from-date
        type: string
      - description: Верхняя граница даты (YYYY-MM-DD)
        in: query
        name: to-date
        type: string
      - description: Статус расчета (draft, formed, moderated, deleted)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список расчетов
          schema:
            items:
              $ref: '#/definitions/apitypes.MassCalculationJSON'
            type: array
        "400":
          description: Неверный формат даты или параметров запроса"}
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдены записи
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список расчетов
      tags:
      - mass-calculations
  /mass-calculations/{id}:
    delete:
      description: Помечает расчет как удалённый. Доступ — владелец или модератор.
        Возвращает {"message":"Calculation deleted"}.
      parameters:
      - description: ID расчета
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить расчет (логическое удаление)
      tags:
      - mass-calculations
    get:
      description: Возвращает расчет и список реакций, входящих в неё, с рассчитанными
        массами.
      parameters:
      - description: ID расчета
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Объект расчета с реакциями
          schema:
            $ref: '#/definitions/handler.MassCalculationResponse'
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить расчет по ID
      tags:
      - mass-calculations
    put:
      consumes:
      - application/json
      description: 'Обновляет расчет по ID: принимает JSON расчета и возвращает обновлённый
        объект.'
      parameters:
      - description: ID расчета
        in: path
        name: id
        required: true
        type: integer
      - description: Тело запроса — объект расчета
        in: body
        name: calculation
        required: true
        schema:
          $ref: '#/definitions/apitypes.MassCalculationJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый расчет
          schema:
            $ref: '#/definitions/apitypes.MassCalculationJSON'
        "400":
          description: Неверный формат запроса или тела
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещён
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить поля расчета
      tags:
      - mass-calculations
  /mass-calculations/{id}/form:
    put:
      description: Переводит расчет в статус formed — доступен только создателю.
      parameters:
      - description: ID расчета
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Расчет успешно сформирована
          schema:
            $ref: '#/definitions/apitypes.MassCalculationJSON'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Только создатель может формировать расчет
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать расчет
      tags:
      - mass-calculations
  /mass-calculations/{id}/moderate:
    put:
      consumes:
      - application/json
      description: Устанавливает статус расчета — только модератор может это сделать.
      parameters:
      - description: ID расчета
        in: path
        name: id
        required: true
        type: integer
      - description: Тело запроса с полем status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/apitypes.StatusJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Расчет после модерации
          schema:
            $ref: '#/definitions/apitypes.MassCalculationJSON'
        "400":
          description: Некорректные входные данные"}
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизован"}
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Только модератор может модерацию
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Модерировать расчет (только модератор)
      tags:
      - mass-calculations
  /mass-calculations/calculation-cart:
    get:
      description: Возвращает количество реакций в текущей корзине пользователя и
        ID черновика расчета (если есть).
      produces:
      - application/json
      responses:
        "200":
          description: 'Поля: id (int, -1 если нет черновика), reactions_count (int)'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос (например, неверный токен)"}
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Черновик не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить карточку черновика расчета для пользователя
      tags:
      - mass-calculations
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"